+++ package.json
@@ -25,5 +25,9 @@
   "dependencies": {
     "debug": "^1.0.2"
   },
-  "homepage": "https://github.com/delian/node-sflow"
+  "homepage": "https://github.com/delian/node-sflow",
+  "devDependencies": {
+    "chai": "^1.9.1",
+    "mocha": "^1.20.1"
+  }
 }
