+++ test/packet.test.js
@@ -1,0 +1,75 @@
+
+var expect = require('chai').expect;
+var Packet = require('../lib/packet');
+
+//single flow packet
+var VYOS_PACKET1 = '00000005000000015ae3679c000000000000000400009c400000000100000001000000b00000000200000001000001f40000047800000000000000033fffffff00000003000003e900000010000000000000000000000000000000000226000200000008000000030000000000000001000000600000000100000052000000040000004e80ee73955628001b2fb948490800450005be1adf40007f0602c60a64006aadc22005f0d501bb971fe8c91f40b4d55010010000fc000017030307d9000000000000002cf366f907b39f5eab2b95230000';
+//multi flow packet
+var VYOS_PACKET2 = '00000005000000015ae3679c0000000000001d7b00053bd80000000700000001000000980000e2a800000001000000010000e2a800000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb9484908004500002808be40007f061a740a640054adc220249f8601bb71f0016d942d9ca05010010090c90000474500000001000000b00000e2a900000001000000010000e2a900000000000000033fffffff00000003000003e900000010000000000000000000000000000000000226000200000008000000030000000000000001000000600000000100000052000000040000004e80ee73955628001b2fb9484908004500005508bf40007f061a460a640054adc220249f8601bb71f0016d942d9ca050180100b7eb000017030300280000000000000007f755a63c0c43a57c53e847474500000001000000980000e2aa00000001000000010000e2aa00000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb9484908004500002808c040007f061a720a640054adc220249f8601bb71f0019a942da7b050100100858c0000516900000001000000980000e2ab00000001000000010000e2ab00000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb9484908004500002808c140007f061a710a640054adc220249f8601bb71f0019a942db2c0501001007a7c0000adc200000001000000980000e2ac00000001000000010000e2ac00000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb9484908004500002808c240007f061a700a640054adc220249f8601bb71f0019a942dbdd0501001006f6c0000170300000001000000b00000e2ad00000001000000010000e2ad00000000000000033fffffff00000003000003e900000010000000000000000000000000000000000226000200000008000000030000000000000001000000600000000100000052000000040000004e80ee73955628001b2fb948490800450000460b2f40007f06f0990a640046adc2477d122414663d35f97ddb7d0cbe5018f84c2a91000017030300196b7b89864c5e12436f79e22be37cd0b02e29b7adc200000001000000980000e2ae00000001000000010000e2ae00000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb9484908004500002808c340007f061a6f0a640054adc220249f8601bb71f0019a942dc8e050100100645c0000adc2';
+//needs readCounterRecords
+var VYOS_PACKET3 = '00000005000000015ae3679c00000000000002380000465000000008000000020000006c000000010000000100000001000000010000005800000003000000060000000005f5e1000000000100000003000000000003d4b80000117c0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000b00000117d00000001000000010000117d00000000000000033fffffff00000003000003e900000010000000000000000000000000000000000226000200000008000000030000000000000001000000600000000100000052000000040000004e80ee73955628001b2fb94849080045000055530f40007f062d7e0a64004e1f0d5157e1af01bbb4c0559be5c54fa85018069850300000170303002800000000000001c35a4cde958274a3559a89befd9a00000001000000980000117e00000001000000010000117e00000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb94849080045000028531040007f062daa0a64004e1f0d5157e1af01bbb4c055c8e5c559805010069800ed0000484900000001000000980000117f00000001000000010000117f00000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb94849080045000028531140007f062da90a64004e1f0d5157e1af01bbb4c055c8e5c5635850100698f7140000484900000001000000980000118000000001000000010000118000000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb94849080045000028531240007f062da80a64004e1f0d5157e1af01bbb4c055c8e5c56d3050100698ed3c0000484900000001000000980000118100000001000000010000118100000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb94849080045000028531340007f062da70a64004e1f0d5157e1af01bbb4c055c8e5c5770850100698e3640000484900000001000000980000118200000001000000010000118200000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb94849080045000028531440007f062da60a64004e1f0d5157e1af01bbb4c055c8e5c580e050100693d9910000484900000001000000980000118300000001000000010000118300000000000000033fffffff00000003000003e90000001000000000000000000000000000000000022600020000000800000003000000000000000100000048000000010000003a000000040000003680ee73955628001b2fb94849080045000028531540007f062da50a64004e1f0d5157e1af01bbb4c055c8e5c58ab850100693cfb900004849';
+
+
+
+describe('Packet', function () {
+
+    it('should decode vyos', function (done) {
+        var p = new Packet(new Buffer(VYOS_PACKET1, 'hex'));
+        p.on('done', function() {
+            expect(p).to.have.property('header');
+            expect(p).to.have.property('flows');
+            expect(p.flows).to.have.length(1);
+
+            var flow = p.flows[0];        
+            expect(flow).to.have.property('input', 3);
+            expect(flow).to.have.property('records');
+            var records = flow.records;
+            expect(records).to.have.length(flow.input);
+
+            records.forEach(function (r) {
+                if(r.enterprise === 8800 && r.format === 2) {
+                    expect(r).to.have.property('tag', 3);
+                    expect(r).to.have.property('tag2', 0);
+                }
+            });    
+            
+            done();    
+        });
+    });
+
+
+    it('should decode vyos and emit on each flow', function (done) {
+        var p = new Packet(new Buffer(VYOS_PACKET2, 'hex'));
+        var count = 0;
+        p.on('flow', function(flow) {
+            count +=1;
+            expect(flow).to.have.property('format', 1);
+            expect(flow).to.have.property('input', 3);
+            expect(flow).to.have.property('records');
+        });
+        
+        setTimeout(function () {
+            expect(count).to.be.equal(7);
+            done();
+        }, 100); 
+    });
+
+    it('should support flow.format === 2 //readCounterRecords', function (done) {
+        var p = new Packet(new Buffer(VYOS_PACKET3, 'hex'));
+        p.on('done', function() {
+            expect(p).to.have.property('header');
+            
+            var header = p.header;
+            expect(header).to.have.property('sequence', 568);
+            expect(header).to.have.property('samples', 8);
+
+            expect(p).to.have.property('flows');
+            expect(p.flows).to.have.length(header.samples);
+            
+            var flow = p.flows[0];
+            expect(flow).to.have.property('format', 2);
+
+            done();    
+        });
+    })
+});
